Extract TravelData interface
